back/requirements.txt
asgiref==3.4.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.10
django-admin-tools==0.9.2
django-cors-headers==3.10.1
django-filter==1.1.0
django-templated-mail==1.1.1
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extensions==0.7.1
drf-writable-nested==0.6.3
enum-compat==0.0.3
envparse==0.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
null==0.6.1
oauthlib==3.1.1
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
